Change project folder structure
