--- client/src/sample_data.js
@@ -1,10 +1,0 @@
-const sampleBooking = {
-  booking_id: 1,
-  home_id: 1,
-  user_id: 1,
-  created_at: '2019-02-07 17:28:09',
-  check_in: '2019-03-01',
-  check_out: '2019-03-04',
-  price_per_night: 166,
-  no_guests: 2
-};
