--- npm-debug.log
@@ -1,48 +1,0 @@
-0 info it worked if it ends with ok
-1 verbose cli [ '/Users/D3/.nvm/versions/node/v6.16.0/bin/node',
-1 verbose cli   '/Users/D3/.nvm/versions/node/v6.16.0/bin/npm',
-1 verbose cli   'run',
-1 verbose cli   'test' ]
-2 info using npm@3.10.10
-3 info using node@v6.16.0
-4 verbose run-script [ 'pretest', 'test', 'posttest' ]
-5 info lifecycle booking-module@1.0.0~pretest: booking-module@1.0.0
-6 silly lifecycle booking-module@1.0.0~pretest: no script for pretest, continuing
-7 info lifecycle booking-module@1.0.0~test: booking-module@1.0.0
-8 verbose lifecycle booking-module@1.0.0~test: unsafe-perm in lifecycle true
-9 verbose lifecycle booking-module@1.0.0~test: PATH: /Users/D3/.nvm/versions/node/v6.16.0/lib/node_modules/npm/bin/node-gyp-bin:/Users/D3/booking-module/node_modules/.bin:/usr/local/opt/mysql@5.7/bin:/Users/D3/.nvm/versions/node/v6.16.0/bin:/Users/D3/anaconda/bin:/usr/bin:/bin:/usr/sbin:/sbin:/usr/local/bin:/Applications/Postgres.app/Contents/Versions/latest/bin
-10 verbose lifecycle booking-module@1.0.0~test: CWD: /Users/D3/booking-module
-11 silly lifecycle booking-module@1.0.0~test: Args: [ '-c', 'jest' ]
-12 silly lifecycle booking-module@1.0.0~test: Returned: code: 1  signal: null
-13 info lifecycle booking-module@1.0.0~test: Failed to exec test script
-14 verbose stack Error: booking-module@1.0.0 test: `jest`
-14 verbose stack Exit status 1
-14 verbose stack     at EventEmitter.<anonymous> (/Users/D3/.nvm/versions/node/v6.16.0/lib/node_modules/npm/lib/utils/lifecycle.js:255:16)
-14 verbose stack     at emitTwo (events.js:106:13)
-14 verbose stack     at EventEmitter.emit (events.js:191:7)
-14 verbose stack     at ChildProcess.<anonymous> (/Users/D3/.nvm/versions/node/v6.16.0/lib/node_modules/npm/lib/utils/spawn.js:40:14)
-14 verbose stack     at emitTwo (events.js:106:13)
-14 verbose stack     at ChildProcess.emit (events.js:191:7)
-14 verbose stack     at maybeClose (internal/child_process.js:920:16)
-14 verbose stack     at Process.ChildProcess._handle.onexit (internal/child_process.js:230:5)
-15 verbose pkgid booking-module@1.0.0
-16 verbose cwd /Users/D3/booking-module
-17 error Darwin 17.7.0
-18 error argv "/Users/D3/.nvm/versions/node/v6.16.0/bin/node" "/Users/D3/.nvm/versions/node/v6.16.0/bin/npm" "run" "test"
-19 error node v6.16.0
-20 error npm  v3.10.10
-21 error code ELIFECYCLE
-22 error booking-module@1.0.0 test: `jest`
-22 error Exit status 1
-23 error Failed at the booking-module@1.0.0 test script 'jest'.
-23 error Make sure you have the latest version of node.js and npm installed.
-23 error If you do, this is most likely a problem with the booking-module package,
-23 error not with npm itself.
-23 error Tell the author that this fails on your system:
-23 error     jest
-23 error You can get information on how to open an issue for this project with:
-23 error     npm bugs booking-module
-23 error Or if that isn't available, you can get their info via:
-23 error     npm owner ls booking-module
-23 error There is likely additional logging output above.
-24 verbose exit [ 1, true ]
