+++ public/index.html
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html>
+<head>
+  <title>Booking</title>
+  <link >
+</head>
+<body>
+  <div id="app"></div>
+  <script type="text/javascript" src="./bundle.js"></script>
+</body>
+</html>
