--- tests/BookingWidget.test.js
@@ -1,6 +1,0 @@
-import React from 'react';
-import renderer from 'react-test-renderer';
-
-test('it works', () => {
-  expect(true).toBe(true);
-});
