+++ webpack.config.js
@@ -3,21 +3,26 @@ var SRC_DIR = path.join(__dirname, '/client/src');
 var DIST_DIR = path.join(__dirname, '/client/dist');
 
 module.exports = {
-  entry: `${SRC_DIR}/index.jsx`,
+  entry: `${__dirname}/client/src/index.jsx`,
   output: {
     filename: 'bundle.js',
-    path: DIST_DIR
+    path: `${__dirname}/public`,
   },
   module: {
     rules: [
       {
         test: /\.jsx?/,
         exclude: /node_modules/,
-        include: SRC_DIR,
         use: {
-          loader: 'babel-loader'
-        }
-      }
-    ]
-  }
+          loader: 'babel-loader',
+          options: {
+            presets: ['@babel/preset-react', '@babel/preset-env'],
+          },
+        },
+      },
+    ],
+  },
+  resolve: {
+    extensions: ['.js', '.jsx'],
+  },
 };
